Changed remove
Fixed the remove function removing channels that people are still talking in.

+++ channels.js
@@ -19,9 +19,10 @@ module.exports = {
             break;
                 
             case 'remove':
-                message.guild.channels.findAll('name', 'auto').forEach(function(channel) {
-                    
-                    channel.delete()
+                message.guild.channels.findAll('name', 'auto').forEach(function(channel, item) {
+                    if (item.members.length > 0) {
+                        channel.delete()
+                    }
                 })
                 
             break;
